hackathon: cluster targets query
